+++ main/src/ca/uwaterloo/flix/language/phase/inference/RestrictableChooseInference.scala
@@ -18,7 +18,7 @@ package ca.uwaterloo.flix.language.phase.inference
 import ca.uwaterloo.flix.api.Flix
 import ca.uwaterloo.flix.language.ast.KindedAst.RestrictableChoosePattern
 import ca.uwaterloo.flix.language.ast.{Ast, Kind, KindedAst, Scheme, SourceLocation, Symbol, Type, TypeConstructor}
-import ca.uwaterloo.flix.language.phase.TypeInference.inferExp
+import ca.uwaterloo.flix.language.phase.Typer.inferExp
 import ca.uwaterloo.flix.language.phase.unification.InferMonad
 import ca.uwaterloo.flix.language.phase.unification.InferMonad.traverseM
 import ca.uwaterloo.flix.language.phase.unification.Unification.{expectTypeM, liftM, unifyTypeM}
